Extract the subprocess VCS probes into their own module
lib.rs carried the marker walk and the subprocess revision probe inline while
the library-backed adapter sat in its own file, so the crate root was both a
composition root and one of the two adapters. Move MarkerWalkRoot,
CommandProbe, the environment scrub, the capped spawn and the hand-rolled
.jj/repo pointer read into vcs_adapters::subprocess, mirroring
vcs_adapters::library. lib.rs drops from 315 lines to 59: the module
declarations, facts, and the one test that covers the composition root. The
unit tests move with the code they exercise.

The layout now says what pup.ron already enforced — spawning is confined to
one module, which is what lets the sibling carry a rule denying std::process.
Retiring the subprocess pair becomes a file deletion, and markers.rs survives
it because both adapters delegate to that rather than to each other; recorded
on 0185, which owns the deletion.

Two consequences for the pin rule's comment: it named crate::CommandProbe,
which is now crate::subprocess. And it had absorbed the config command rule's
comment, orphaned nine lines above the wrong Module block when the VCS rule
was first inserted — that explanation of why it uses `denied` rather than
`allowed_only` is back with its own rule.
