+++ cli/vcs-adapters/src/subprocess.rs
@@ -1,0 +1,306 @@
+//! The subprocess-backed probes: an ancestor marker-walk for the repository
+//! root, and the VCS binaries themselves for the working-copy revision.
+//!
+//! The probe subprocess runs with a scrubbed environment and colour disabled,
+//! so ambient user config cannot redirect the root or inject ANSI into a
+//! revision. Every way the probe can fail to answer — an absent binary, a
+//! non-zero exit, empty output, or a run that outlives its time cap — resolves
+//! to `None` and is warn-logged, so a real failure leaves a trace instead of
+//! reading like a legitimately revision-less repository.
+//!
+//! This is the pair [`crate::library`] replaces. The two differ in more than
+//! mechanism: this one parses in a child process with a time cap, a
+//! kill-on-timeout and a scrubbed environment, which is containment the
+//! in-process reader cannot offer. Whether an equivalent bound is needed before
+//! the composition root switches over is a decision the switch owns.
+//!
+//! Spawning lives here and only here, and this module is deliberately unruled
+//! where [`crate::library`] carries a cargo-pup rule denying `std::process`.
+//! That rule is module-scoped, so it did not need this extraction to work — what
+//! the extraction buys is that the boundary is now visible in the file layout
+//! rather than only in `pup.ron`, and that retiring this pair is a file
+//! deletion.
+
+use std::fs;
+use std::io::Read as _;
+use std::path::{Path, PathBuf};
+use std::process::{Command, Stdio};
+use std::time::{Duration, Instant};
+
+use tracing::warn;
+use vcs::{RepoRoot, VcsKind, VcsProbe};
+
+use crate::markers::{carries_any_marker, marker_kind, walk_up};
+
+/// Headroom for a legitimately slow or lock-contended repository, while still
+/// bounding metadata derivation.
+const DEFAULT_CAP: Duration = Duration::from_secs(10);
+
+const POLL_INTERVAL: Duration = Duration::from_millis(10);
+
+/// Finds the repository root by walking ancestors for the first `.jj` or
+/// `.git` marker, testing for *existence* so a `.git` file — a worktree or
+/// submodule — counts alongside a `.git` directory.
+///
+/// The filesystem root itself is never tested, matching the bash walk.
+#[derive(Debug, Clone, Copy, Default)]
+pub struct MarkerWalkRoot;
+
+impl RepoRoot for MarkerWalkRoot {
+    fn discover(&self, start: &Path) -> Option<PathBuf> {
+        walk_up(start, carries_any_marker)
+    }
+
+    fn repository_root(&self, working_copy_root: &Path) -> PathBuf {
+        jj_repository_root(working_copy_root)
+            .unwrap_or_else(|| working_copy_root.to_path_buf())
+    }
+}
+
+/// Resolves a jj secondary workspace's working-copy root to the repository whose
+/// store it shares. A secondary workspace's `.jj/repo` is a file pointing at the
+/// shared `<repo>/.jj/repo` store, so the repository is the store's grandparent;
+/// the primary workspace's `.jj/repo` is that store directory. `None` when this
+/// is not a jj workspace or the pointer cannot be resolved.
+///
+/// This reads the pointer by hand. [`crate::library`] resolves the same fact
+/// through jj-lib's workspace loader instead, so the file-versus-directory rule
+/// has one implementation per adapter rather than one shared one — the loader
+/// route needs the library this module deliberately does not depend on.
+fn jj_repository_root(working_copy_root: &Path) -> Option<PathBuf> {
+    let marker = working_copy_root.join(".jj").join("repo");
+    let metadata = fs::symlink_metadata(&marker).ok()?;
+    if !metadata.is_file() {
+        return None;
+    }
+    let pointer = fs::read_to_string(&marker).ok()?;
+    let store =
+        fs::canonicalize(working_copy_root.join(".jj").join(pointer.trim()))
+            .ok()?;
+    Some(store.parent()?.parent()?.to_path_buf())
+}
+
+/// Reads the repository's idiom from its markers and its revision by running
+/// the matching VCS binary.
+#[derive(Debug, Clone, Copy)]
+pub struct CommandProbe {
+    cap: Duration,
+}
+
+impl CommandProbe {
+    #[must_use]
+    pub const fn new() -> Self {
+        Self { cap: DEFAULT_CAP }
+    }
+
+    /// A probe bounded by `cap` rather than the default.
+    #[must_use]
+    pub const fn with_cap(cap: Duration) -> Self {
+        Self { cap }
+    }
+}
+
+impl Default for CommandProbe {
+    fn default() -> Self {
+        Self::new()
+    }
+}
+
+impl VcsProbe for CommandProbe {
+    fn kind(&self, root: &Path) -> VcsKind {
+        marker_kind(root)
+    }
+
+    fn revision(&self, root: &Path, kind: VcsKind) -> Option<String> {
+        let mut command = match kind {
+            VcsKind::Jj => {
+                let mut command = Command::new("jj");
+                command.args([
+                    "--color=never",
+                    "--no-pager",
+                    "log",
+                    "-r",
+                    "@",
+                    "--no-graph",
+                    "-T",
+                    "commit_id",
+                ]);
+                command
+            }
+            VcsKind::Git => {
+                let mut command = Command::new("git");
+                command.args(["-c", "color.ui=false", "rev-parse", "HEAD"]);
+                command
+            }
+            VcsKind::None => return None,
+        };
+
+        command.current_dir(root);
+        scrub_environment(&mut command);
+        capped_stdout(command, self.cap, kind.as_str())
+    }
+}
+
+/// Denies the subprocess the ambient config that could redirect it at another
+/// repository or dress its output up in ANSI.
+fn scrub_environment(command: &mut Command) {
+    for key in [
+        "GIT_DIR",
+        "GIT_WORK_TREE",
+        "GIT_INDEX_FILE",
+        "GIT_COMMON_DIR",
+        "GIT_CONFIG",
+        "GIT_CONFIG_COUNT",
+        "JJ_CONFIG",
+    ] {
+        command.env_remove(key);
+    }
+    command.env("GIT_CONFIG_NOSYSTEM", "1");
+    command.env("GIT_CONFIG_GLOBAL", "/dev/null");
+    command.env("GIT_CONFIG_SYSTEM", "/dev/null");
+}
+
+/// Runs `command` and returns its trimmed stdout, or `None` — warn-logged —
+/// when it cannot be spawned, exits non-zero, outlives `cap`, or says nothing.
+fn capped_stdout(
+    mut command: Command,
+    cap: Duration,
+    vcs: &str,
+) -> Option<String> {
+    command
+        .stdin(Stdio::null())
+        .stdout(Stdio::piped())
+        .stderr(Stdio::piped());
+
+    let mut child = match command.spawn() {
+        Ok(child) => child,
+        Err(error) => {
+            warn!(vcs, %error, "could not run the revision probe");
+            return None;
+        }
+    };
+
+    let deadline = Instant::now() + cap;
+    let status = loop {
+        match child.try_wait() {
+            Ok(Some(status)) => break status,
+            Ok(None) => {}
+            Err(error) => {
+                warn!(vcs, %error, "could not await the revision probe");
+                return None;
+            }
+        }
+
+        if Instant::now() >= deadline {
+            let _ = child.kill();
+            let _ = child.wait();
+            warn!(
+                vcs,
+                cap = ?cap,
+                "the revision probe outlived its time cap and was killed"
+            );
+            return None;
+        }
+
+        std::thread::sleep(POLL_INTERVAL);
+    };
+
+    if !status.success() {
+        warn!(vcs, %status, "the revision probe exited non-zero");
+        return None;
+    }
+
+    let mut stdout = String::new();
+    if let Some(mut pipe) = child.stdout.take() {
+        if let Err(error) = pipe.read_to_string(&mut stdout) {
+            warn!(vcs, %error, "could not read the revision probe's output");
+            return None;
+        }
+    }
+
+    let revision = stdout.trim();
+    if revision.is_empty() {
+        warn!(vcs, "the revision probe reported no revision");
+        return None;
+    }
+    Some(revision.to_owned())
+}
+
+#[cfg(test)]
+mod tests {
+    use std::process::Command;
+    use std::time::Duration;
+
+    use super::{capped_stdout, scrub_environment};
+
+    #[test]
+    fn a_probe_that_outlives_its_cap_reports_no_revision() {
+        let mut command = Command::new("sleep");
+        command.arg("30");
+
+        let started = std::time::Instant::now();
+        let revision =
+            capped_stdout(command, Duration::from_millis(100), "test");
+
+        assert_eq!(revision, None);
+        assert!(
+            started.elapsed() < Duration::from_secs(5),
+            "the probe should have been killed at its cap, not waited out"
+        );
+    }
+
+    #[test]
+    fn a_probe_that_cannot_be_spawned_reports_no_revision() {
+        let command = Command::new("accelerator-no-such-binary");
+        assert_eq!(
+            capped_stdout(command, Duration::from_secs(1), "test"),
+            None
+        );
+    }
+
+    #[test]
+    fn a_probe_that_exits_non_zero_reports_no_revision() {
+        let command = Command::new("false");
+        assert_eq!(
+            capped_stdout(command, Duration::from_secs(1), "test"),
+            None
+        );
+    }
+
+    #[test]
+    fn a_probe_that_says_nothing_reports_no_revision() {
+        let command = Command::new("true");
+        assert_eq!(
+            capped_stdout(command, Duration::from_secs(1), "test"),
+            None
+        );
+    }
+
+    #[test]
+    fn stdout_is_trimmed() {
+        let mut command = Command::new("printf");
+        command.arg("abc123\n");
+        assert_eq!(
+            capped_stdout(command, Duration::from_secs(1), "test").as_deref(),
+            Some("abc123")
+        );
+    }
+
+    #[test]
+    fn the_scrubbed_environment_drops_the_redirecting_variables() {
+        let mut command = Command::new("sh");
+        command
+            .args([
+                "-c",
+                "if [ -n \"$GIT_DIR\" ]; then printf carried; \
+                 else printf unset; fi",
+            ])
+            .env("GIT_DIR", "/somewhere/else/.git");
+        scrub_environment(&mut command);
+
+        assert_eq!(
+            capped_stdout(command, Duration::from_secs(1), "test").as_deref(),
+            Some("unset")
+        );
+    }
+}
